Consolidate scripts
