+++ package.json
@@ -22,12 +22,11 @@
     "deploy": "DEBUG=browserless* ./scripts/deploy.js",
     "deploy-base": "DEBUG=browserless* ./scripts/deploy-base.js",
     "dev": "npm run build && env-cmd -f ./.env.dev node build/index.js",
-    "meta": "./scripts/get-meta.js",
+    "meta": "./scripts/meta.js",
     "start": "node ./build",
     "test": "DEBUG=quiet jest --runInBand --detectOpenHandles",
-    "symlink-chrome": "node ./scripts/symlink-chrome",
     "external": "rimraf debugger/devtools && node ./scripts/install-external-deps.js",
-    "postinstall": "npm run external && npm run symlink-chrome && npm run meta",
+    "postinstall": "./scripts/postinstall.js",
     "snyk-protect": "snyk protect",
     "prettier": "prettier --config .prettierrc --write --loglevel error '{src,functions,scripts}/**/*.{js,ts}'",
     "lint": "eslint src --ext .ts --fix"
